+++ examples/resources/meraki_appliance_dns_local_profile_assignments/resource.tf
@@ -1,0 +1,9 @@
+resource "meraki_appliance_dns_local_profile_assignments" "example" {
+  organization_id = "123456"
+  items = [
+    {
+      network_id = "N_123456"
+      profile_id = "1234"
+    }
+  ]
+}